Move Job and Token to datatypes package
