+++ src/main/java/com/suse/saltstack/netapi/client/SaltStackClient.java
@@ -7,9 +7,9 @@ import static com.suse.saltstack.netapi.config.ClientConfig.*;
 import com.suse.saltstack.netapi.config.ProxySettings;
 import com.suse.saltstack.netapi.exception.SaltStackException;
 import com.suse.saltstack.netapi.parser.JsonParser;
-import com.suse.saltstack.netapi.results.Job;
+import com.suse.saltstack.netapi.datatypes.Job;
 import com.suse.saltstack.netapi.results.Result;
-import com.suse.saltstack.netapi.results.Token;
+import com.suse.saltstack.netapi.datatypes.Token;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
