+++ src/main/java/com/suse/saltstack/netapi/parser/JsonParser.java
@@ -7,6 +7,8 @@ import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParseException;
 import com.google.gson.reflect.TypeToken;
+import com.suse.saltstack.netapi.datatypes.Job;
+import com.suse.saltstack.netapi.datatypes.Token;
 import com.suse.saltstack.netapi.results.*;
 
 import java.io.BufferedReader;
